fix : implementing dto's and fix errors+ add feedback errors
Fix errors, add DTOs for panne and onderhoud creation requests, adjust addCategorie function in CategorieService etc

+++ rollecoaster/src/main/kotlin/be/ehb/rollecoaster/dto/PanneRequest.kt
@@ -2,4 +2,4 @@ package be.ehb.rollecoaster.dto
 
 import java.util.Date
 
-data class PanneRequest(val attractieId:Int,val description:String,val beginDate: Date, val endDate: Date)
+data class PanneRequest(val attractieId:Long,val description:String,val beginDate: Date, val endDate: Date)
